requirements.txt
streamlit==1.32.2
openai==2.8.1
python-dotenv==1.0.1
